+++ Utility/DataSource.cpp
@@ -2,7 +2,7 @@
 
 namespace mobo
 {
-    void DataSource::resizeIfNeeded(uint32_t iNewSize, bool iPreserve)
+    void DataSource::resizeIfNeeded(size_t iNewSize, bool iPreserve)
     {
         if(iNewSize != size()) {
             setSize(iNewSize, iPreserve);
